+++ test/e2e/templates/basic-dev.tpl
@@ -1,6 +1,8 @@
 {{#> devLayout}}
 var options = {
-  'baseUrl': '{{{WIDGET_TEST_SERVER}}}'
+  'baseUrl': '{{{WIDGET_TEST_SERVER}}}',
+  'el': '#okta-login-container'
+
 };
 var oktaSignIn = new OktaSignIn(options);
 {{/devLayout}}
